chore: fix lint issues
