+++ .eslintrc.js
@@ -17,6 +17,8 @@ module.exports = {
   env: {
     browser: true,
     node: true,
+    es6: true,
+    jest: true,
   },
   rules: {
     // I'll probably add some typescript types instead
