+++ postcss.config.js
@@ -10,7 +10,7 @@ if (!IS_DEVELOPMENT) {
 
   class TailwindExtractor {
     static extract(content) {
-      return content.match(/[A-z0-9-:\/]+/g) || [];
+      return content.match(/[A-z0-9-:/]+/g) || [];
     }
   }
